+++ libs/core/product-switch/product-switch-body/product-switch-body.component.ts
@@ -17,6 +17,7 @@ import { Subscription } from 'rxjs';
 import { FdDropEvent, KeyUtil, RtlService } from '@fundamental-ngx/cdk/utils';
 
 import { DndItemDirective, DndListDirective } from '@fundamental-ngx/cdk/utils';
+import { AvatarComponent } from '@fundamental-ngx/core/avatar';
 import { FD_DEFAULT_ICON_FONT_FAMILY, IconComponent } from '@fundamental-ngx/core/icon';
 import { ProductSwitchItem } from './product-switch.item';
 
@@ -29,7 +30,7 @@ const containerWidthPx = 776;
     styleUrls: ['./product-switch-body.component.scss', '../../../cdk/utils/drag-and-drop/drag-and-drop.scss'],
     encapsulation: ViewEncapsulation.None,
     changeDetection: ChangeDetectionStrategy.OnPush,
-    imports: [DndItemDirective, DndListDirective, IconComponent]
+    imports: [DndItemDirective, DndListDirective, IconComponent, AvatarComponent]
 })
 export class ProductSwitchBodyComponent implements OnInit, OnDestroy {
     /** Defines if drag and drop functionality should be included in product switch*/