+++ libs/docs/cdk/disabled/disabled-header/disabled-header.component.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, signal } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
 import {
     DescriptionComponent,
@@ -11,4 +11,6 @@ import {
     templateUrl: './disabled-header.component.html',
     imports: [HeaderComponent, DescriptionComponent, ImportComponent, HeaderTabsComponent, RouterOutlet]
 })
-export class DisabledHeaderComponent {}
+export class DisabledHeaderComponent {
+    falseSignal = signal(false);
+}