+++ libs/docs/cdk/forms/forms-header/forms-header.component.ts
@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component, signal } from '@angular/core';
 import { RouterLink } from '@angular/router';
 import {
     DescriptionComponent,
@@ -14,4 +14,6 @@ import {
     changeDetection: ChangeDetectionStrategy.OnPush,
     imports: [DocPageComponent, HeaderComponent, DescriptionComponent, RouterLink, ImportComponent, HeaderTabsComponent]
 })
-export class FormsHeaderComponent {}
+export class FormsHeaderComponent {
+    falseSignal = signal(false);
+}