+++ libs/docs/core/icon/examples/icon-example.component.ts
@@ -706,6 +706,8 @@ export class IconExampleComponent {
         'in-progress-2',
         'unsynchronize',
         'verified',
-        'walk-me'
+        'walk-me',
+        'smart-watch',
+        'vr-glasses'
     ];
 }