+++ libs/i18n/src/lib/translations/translations_en_US_sappsd.ts
@@ -71,7 +71,12 @@ export default {
         listItemStatusDraft: '[[[Ďŗąƒţ∙∙∙∙∙∙∙∙∙]]]'
     },
     coreMessageStrip: {
-        dismissLabel: '[[[Ďįşɱįşş∙∙∙∙∙∙∙]]]'
+        dismissLabel: '[[[Ďįşɱįşş∙∙∙∙∙∙∙]]]',
+        announcementInfo: 'Information Message Strip',
+        announcementWarning: 'Warning Message Strip',
+        announcementError: 'Error Message Strip',
+        announcementSuccess: 'Success Message Strip',
+        closable: 'Closable'
     },
     coreMultiInput: {
         countListResultsSingular: '[[[1 ŗēşűĺţ ĺįşţ įţēɱ.∙∙∙∙∙]]]',