+++ libs/platform/approval-flow/approval-flow-user-list/approval-flow-user-list.component.ts
@@ -3,7 +3,9 @@ import {
     ChangeDetectionStrategy,
     ChangeDetectorRef,
     Component,
+    DestroyRef,
     EventEmitter,
+    inject,
     Input,
     OnChanges,
     OnDestroy,
@@ -15,6 +17,7 @@ import {
     ViewEncapsulation
 } from '@angular/core';
 
+import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { FormsModule } from '@angular/forms';
 import { FdTranslatePipe } from '@fundamental-ngx/i18n';
 import {
@@ -88,7 +91,16 @@ export class ApprovalFlowUserListComponent implements AfterViewInit, OnChanges, 
     _displayUsers: ApprovalUser[] = [];
 
     /** @hidden */
+    _allSelectedUsers: ApprovalUser[] = [];
+
+    /** @hidden */
     private _intervalID?: number;
+
+    /** @hidden */
+    private _destroyed$ = inject(DestroyRef);
+
+    /** @hidden */
+    private _dontCallSelection = false;
 
     /** @hidden */
     constructor(private _cdr: ChangeDetectorRef) {}
@@ -96,18 +108,13 @@ export class ApprovalFlowUserListComponent implements AfterViewInit, OnChanges, 
     /** @hidden */
     ngAfterViewInit(): void {
         if (this.selectedUsers.length) {
-            const selectedApproversNames = this.selectedUsers.map((approver) => approver.name);
+            this._selectItems(this.selectedUsers);
+            this._allSelectedUsers = [...this.selectedUsers];
+            this._cdr.markForCheck();
+        }
 
-            this._selectedItems = this.listItems.filter(
-                (item) => !!item.avatar?.ariaLabel && selectedApproversNames.includes(item.avatar.ariaLabel)
-            );
-
-            this._selectedItems.forEach((item) => {
-                item._selected = true;
-                this.list._selectItem(item);
-            });
-
-            this._cdr.detectChanges();
+        if (this.isSelectable) {
+            this._listenToListItemChange();
         }
     }
 
@@ -125,15 +132,29 @@ export class ApprovalFlowUserListComponent implements AfterViewInit, OnChanges, 
 
     /** @hidden */
     _onSelect(event: SelectionChangeEvent): void {
-        this._selectedItems = event.selectedItems;
+        if (!this._dontCallSelection) {
+            this._selectedItems = event.selectedItems;
 
-        this.onSelectionChange.emit(this._getUsersFromSelectedItems(event.selectedItems));
+            if (event.removed) {
+                const ids = event.removed.map((item: BaseListItem) => item.id);
+                this._allSelectedUsers = this._allSelectedUsers.filter(
+                    (user: ApprovalUser) => !ids.includes(`${this._idPrefix + user.id}`)
+                );
+            }
+
+            if (event.added) {
+                this._allSelectedUsers.push(...this._getUsersFromSelectedItems(event.added));
+            }
+
+            this.onSelectionChange.emit(this._allSelectedUsers);
+            this._cdr.detectChanges();
+        }
     }
 
     /** @hidden */
     private _getUsersFromSelectedItems(items: BaseListItem[]): ApprovalUser[] {
         return items
-            .map((item) => this.users.find((user) => `${this._idPrefix + user.id}` === item.itemEl.nativeElement.id))
+            .map((item) => this.users.find((user) => `${this._idPrefix + user.id}` === item.id))
             .filter((u): u is ApprovalUser => !!u);
     }
 
@@ -182,5 +203,27 @@ export class ApprovalFlowUserListComponent implements AfterViewInit, OnChanges, 
         if (this._intervalID) {
             clearInterval(this._intervalID);
         }
+    }
+
+    /** @hidden */
+    private _listenToListItemChange(): void {
+        this.listItems.changes.pipe(takeUntilDestroyed(this._destroyed$)).subscribe(() => {
+            this._selectItems(this._allSelectedUsers);
+            this._cdr.detectChanges();
+        });
+    }
+
+    /** @hidden */
+    private _selectItems(users: ApprovalUser[]): void {
+        this._dontCallSelection = true;
+        const allSelectedUserNames = users.map((user) => this._idPrefix + user.id);
+        const currentSelectedList = this.listItems.filter((item) => allSelectedUserNames.includes(item.id));
+        this.list._clearSelection();
+        this._selectedItems = [...currentSelectedList];
+        this._selectedItems.forEach((item) => {
+            item._selected = true;
+            this.list._selectItem(item);
+        });
+        this._dontCallSelection = false;
     }
 }